+++ docs/Changelog.md
@@ -112,6 +112,7 @@
 * [#7614](https://github.com/TouK/nussknacker/pull/7614) SpelTemplate as a main text editor
   * Added a new parameter editor type: SpelParameterEditor which works the same as RawParameterEditor.
   * Added a list of editors to the UIParameter.
+  * Added migration of scenario nodes from a string defined in SpEL to a SpelTemplate string. 
 * [#7616](https://github.com/TouK/nussknacker/pull/7616) (K8s DM) Fix for: k8s object name sanitizing strategy sometimes generated invalid object names, in other cases, 
   it generated names with unnecessary characters appended
 * [#7615](https://github.com/TouK/nussknacker/pull/7615) Updated Flink dependency to 1.19.2